requirements.txt
alembic==1.5.8
aniso8601==9.0.1
APScheduler==3.6.3
argcomplete==1.12.2
argon2-cffi==20.1.0
arrow==1.0.3
asgiref==3.3.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifultable==1.0.0
binaryornot==0.4.4
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
emoji==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
graphviz==0.14.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
kiwisolver==1.2.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
matplotlib==3.3.2
mistune==0.8.4
mlxtend==0.17.3
nbclient==0.5.0
nbconvert==6.0.5
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.5
notebook==6.1.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
pipx==0.16.1.0
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.1
Pygments==2.7.1
pymongo==3.11.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
python-telegram-bot==13.4.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0 
sqlparse==0.4.1
tabulate==0.8.7
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
userpath==1.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.2.0